+++ client/src/pages/Account/AccountContent/AccountContent.tsx
@@ -4,10 +4,8 @@ import AccountNotifications from '../AccountNotifications/AccountNotifications';
 import { AccountProfile } from '../AccountProfile/AccountProfile';
 import AccountSocial from '../AccountSocial/AccountSocial';
 import { AccountLocation } from '../contexts/AccountLocationContext';
-import AccountWishlists from '../AccountWishlists/AccountWishlists';
 import AccountProfileProvider from '../providers/AccountProfileProvider';
 import AccountSocialProvider from '../providers/AccountSocialProvider';
-import AccountSocialDetailsProvider from '../providers/AccountSocialDetailsProvider';
 
 export default function AccountContent(): JSX.Element {
   const { accountLocation } = useAccountLocation();
@@ -33,5 +31,4 @@ const contentRecord: Record<AccountLocation, JSX.Element> = {
   ),
 
   notifications: <AccountNotifications />,
-  wishlists: <AccountWishlists />,
 };