+++ client/src/pages/AccountVerification/components/ConfirmAccountVerification.tsx
@@ -1,21 +1,26 @@
-import { JSX, useCallback, useEffect, useState } from 'react';
+import { Dispatch, JSX, SetStateAction, useCallback, useEffect, useState } from 'react';
 import useAuth from '../../../hooks/useAuth';
 import useHandleAsyncError, { HandleAsyncErrorFunction } from '../../../hooks/useHandleAsyncError';
 import { NavigateFunction, useNavigate } from 'react-router-dom';
 import usePopupMessage from '../../../hooks/usePopupMessage';
-import Button from '../../../components/Button/Button';
 import { verifyAccountService } from '../../../services/accountServices';
 import { CanceledError } from 'axios';
+import InstructionCard from '../../../components/InstructionCard/InstructionCard';
 
 type ConfirmAccountVerificationProps = {
   publicAccountId: string;
   verificationToken: string;
+  setIsValidVerificationLink: Dispatch<SetStateAction<boolean>>;
 };
 
-export default function ConfirmAccountVerification({ publicAccountId, verificationToken }: ConfirmAccountVerificationProps): JSX.Element {
+export default function ConfirmAccountVerification({
+  publicAccountId,
+  verificationToken,
+  setIsValidVerificationLink,
+}: ConfirmAccountVerificationProps): JSX.Element {
   const [verificationFailed, setVerificationFailed] = useState<boolean>(false);
   const [title, setTitle] = useState<string>('Ongoing account verification detected.');
-  const [description, setDescription] = useState<string>('Find the verification email in your inbox, and click the link to continue.');
+  const [description, setDescription] = useState<string>('Check your inbox for the verification email and click the link to continue.');
   const [btnTitle, setBtnTitle] = useState<string>('Resend email');
   const [btnNavigateLocation, setBtnNavigateLocation] = useState<string | null>(null);
 
@@ -27,10 +32,17 @@ export default function ConfirmAccountVerification({ publicAccountId, verificati
   const verifyAccount = useCallback(
     async (abortSignal: AbortSignal = new AbortController().signal) => {
       try {
-        await verifyAccountService({ publicAccountId, verificationToken }, abortSignal);
+        const authSessionCreated: boolean = (await verifyAccountService({ publicAccountId, verificationToken }, abortSignal)).data
+          .authSessionCreated;
 
         displayPopupMessage('Account verified.', 'success');
-        navigate('/sign-in');
+
+        if (!authSessionCreated) {
+          navigate('/sign-in');
+          return;
+        }
+        setAuthStatus('authenticated');
+        navigate('/account');
       } catch (err: unknown) {
         if (err instanceof CanceledError) {
           return;
@@ -39,13 +51,15 @@ export default function ConfirmAccountVerification({ publicAccountId, verificati
         console.log(err);
         const { isHandled, status, errMessage, errReason } = handleAsyncError(err);
 
+        setVerificationFailed(true);
+
         if (status === 500) {
-          setVerificationFailed(true);
           displayPopupMessage('Something went wrong.', 'error');
 
           setTitle('Something went wrong.');
           setDescription('Account verification failed due to an unexpected error.');
           setBtnTitle('Try again');
+          setBtnNavigateLocation(null);
 
           return;
         }
@@ -54,39 +68,31 @@ export default function ConfirmAccountVerification({ publicAccountId, verificati
           return;
         }
 
+        if (status === 400) {
+          displayPopupMessage('Invalid recovery link.', 'error');
+          setIsValidVerificationLink(false);
+
+          return;
+        }
+
         if (status === 403) {
           setAuthStatus('authenticated');
           return;
         }
 
-        setVerificationFailed(true);
         setTitle(errMessage);
 
         if (status === 401) {
-          const accountDeleted: boolean = errReason?.includes('_deleted') === true;
-          const description: string = accountDeleted
-            ? 'Account deleted due to too many failed verification attempts.'
-            : `Your verification link is invalid or malformed. Make sure you've copied the correct link.`;
+          const accountDeleted: boolean = errReason === 'incorrectVerificationToken_deleted';
 
-          setDescription(description);
-          setBtnTitle(accountDeleted ? 'Sign up again' : 'Go to homepage');
-          setBtnNavigateLocation('/sign-up');
+          setDescription(
+            accountDeleted
+              ? 'Account deleted due to too many failed verification attempts. You can start over any time.'
+              : `Please make sure you use the full recovery link we've sent you.`
+          );
 
-          return;
-        }
-
-        if (status === 400) {
-          setDescription(`Your verification link is invalid or malformed. Make sure you've copied the correct link.`);
           setBtnTitle('Go to homepage');
           setBtnNavigateLocation('/home');
-
-          return;
-        }
-
-        if (status === 404) {
-          setDescription(`Account doesn't exist or may have been deleted after remaining unverified for longer than 20 minutes.`);
-          setBtnTitle('Sign up');
-          setBtnNavigateLocation('/sign-up');
 
           return;
         }
@@ -95,10 +101,20 @@ export default function ConfirmAccountVerification({ publicAccountId, verificati
           setDescription('You may proceed with signing in.');
           setBtnTitle('Sign in');
           setBtnNavigateLocation('/sign-in');
+
+          return;
+        }
+
+        if (status === 404) {
+          setDescription(`Account either doesn't exist or has had its verification window expire.`);
+          setBtnTitle('Sign up');
+          setBtnNavigateLocation('/sign-up');
+
+          return;
         }
       }
     },
-    [publicAccountId, verificationToken, displayPopupMessage, navigate, setAuthStatus, handleAsyncError]
+    [publicAccountId, verificationToken, setIsValidVerificationLink, displayPopupMessage, navigate, setAuthStatus, handleAsyncError]
   );
 
   useEffect(() => {
@@ -108,32 +124,25 @@ export default function ConfirmAccountVerification({ publicAccountId, verificati
     return () => abortController.abort();
   }, [verifyAccount]);
 
+  if (!verificationFailed) {
+    return <div className='spinner w-3 h-3 mx-auto block'></div>;
+  }
+
   return (
-    <>
-      {verificationFailed ? (
-        <>
-          <h4 className='text-title font-medium mb-1'>{title}</h4>
-          <p className='text-description text-sm'>{description}</p>
+    <InstructionCard
+      title={title}
+      description={description}
+      btnTitle={btnTitle}
+      btnDisabled={false}
+      onClick={async () => {
+        if (btnNavigateLocation) {
+          navigate(btnNavigateLocation);
+          return;
+        }
 
-          <Button
-            className='bg-description border-description text-dark w-full mt-2'
-            onClick={async () => {
-              if (btnNavigateLocation) {
-                navigate(btnNavigateLocation);
-                return;
-              }
-
-              setVerificationFailed(false);
-              await verifyAccount();
-              setVerificationFailed(true);
-            }}
-          >
-            {btnTitle}
-          </Button>
-        </>
-      ) : (
-        <div className='spinner w-3 h-3 mx-auto block'></div>
-      )}
-    </>
+        setVerificationFailed(false);
+        await verifyAccount();
+      }}
+    />
   );
 }