+++ client/src/pages/SignUp/SignUp.tsx
@@ -70,8 +70,7 @@ export default function SignUp(): JSX.Element {
       <section className='py-4 h-available flex justify-center items-center'>
         <Container>
           <div className='py-3 px-2 bg-secondary rounded-sm shadow-simple max-w-[36rem] mx-auto'>
-            <h1 className='text-title text-xl 3xs:text-2xl font-bold text-center'>Sign up to Muninnfy</h1>
-            <div className='h-line my-2'></div>
+            <h1 className='text-title text-xl 3xs:text-2xl font-bold mb-[1.6rem]'>Sign up to Muninnfy</h1>
 
             <form
               id='sign-up-form'