+++ src/cron-jobs/webSocketCronJobs.ts
@@ -1,10 +1,13 @@
-import { WEB_SOCKET_INACTIVITY_THRESHOLD } from '../util/constants/webSocketConstants';
+import { minuteMilliseconds } from '../util/constants/globalConstants';
 import { wsMap } from '../webSocket/webSocketServer';
 
-export function destroyStaleWebSocketsCron(currentTimestamp: number): void {
+export function pingWebSocketsCron(currentTimestamp: number): void {
   for (const { ws, pongTimestamp } of wsMap.values()) {
-    if (currentTimestamp - pongTimestamp >= WEB_SOCKET_INACTIVITY_THRESHOLD) {
+    if (currentTimestamp - pongTimestamp >= minuteMilliseconds * 2) {
       ws.close(1000, 'Connection closed due to inactivity.');
+      continue;
     }
+
+    ws.ping();
   }
 }